First step making the wrapper a HOC
