--- imports/plugins/core/ui/client/components/permission/index.js
@@ -1,1 +1,0 @@
-export * from "./permission";
